+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 - [command_meta.NOT] feature: new `/NOT` command, to reverse the return value of `SUCCESS` and `FAILURE`.
   - Example: `/IF NOT has-item? <player> minecraft:apple 16 THEN say You don't have 16 apples. ELSE say You have 16 apples.`
 - [command_advice] features and bug fixes
-  - feature: improve the interactions of `cancellable advices` (`CANCEL_WITH_SUCCESS` and `CANCEL_WITH_FAILUE`) and `non-cancellable advices`. Now the `non-cancellable advice` will not be performed if the target command execution has already been canceled by a `cancellable command advice`.
+  - feature: improve the interactions of `cancellable advices` (`CANCEL_AS_SUCCESS` and `CANCEL_AS_FAILUE`) and `non-cancellable advices`. Now the `non-cancellable advice` will not be performed if the target command execution has already been canceled by a `cancellable command advice`.
   - feature: new command advice type: `CANCEL_IF_ANY_SUCCESS` and `CANCEL_IF_ALL_SUCCESS`.
   - feature: new command advice type: `ON_EXECUTION_CANCELLED`.
   - feature: new command advice type: `ON_EXECUTION_SUCCESS` and `ON_EXECUTION_FAILURE`.